refactoring server api & requests
film popup view take newComment response from addComment request
