+++ src/model/films-model.js
@@ -1,4 +1,5 @@
 import Observable from '../framework/observable.js';
+import { adaptToClient } from '../service/adapter.js';
 import { UpdateType } from '../consts.js';
 
 export default class FilmsModel extends Observable {
@@ -17,7 +18,7 @@ export default class FilmsModel extends Observable {
   async init() {
     try {
       const films = await this.#filmsApiService.films;
-      this.#films = films.map(this.#adaptToClient);
+      this.#films = films.map(adaptToClient);
     } catch(err) {
       this.#films = [];
     }
@@ -27,54 +28,29 @@ export default class FilmsModel extends Observable {
 
 
   async updateFilm(updateType, update) {
-    const index = this.#films.findIndex((film) => film.id === update.id);
-
-    if (index === -1) {
-      throw new Error('Can\'t update unexisting film');
-    }
-
     try {
       const response = await this.#filmsApiService.updateFilm(update);
-      const updatedFilm = this.#adaptToClient(response);
-      this.#films = [
-        ...this.#films.slice(0, index),
-        updatedFilm,
-        ...this.#films.slice(index + 1),
-      ];
-      this._notify(updateType, updatedFilm);
+      const updatedFilm = adaptToClient(response);
+      this.updateFilmOnClient(updateType, updatedFilm);
     } catch(err) {
       throw new Error('Can\'t update film');
     }
   }
 
-  #adaptToClient(film) {
-    const adaptedFilm = {
-      id: film['id'],
-      comments: film['comments'],
-      filmInfo: {
-        ...film['film_info'],
-        ageRating: film['film_info']['age_rating'],
-        alternativeTitle: film['film_info']['alternative_title'],
-        totalRating: film['film_info']['total_rating'],
-        release: {
-          date: film['film_info']['release']['date'],
-          releaseCountry: film['film_info']['release']['release_country']
-        }
-      },
-      userDetails: {
-        ...film['user_details'],
-        alreadyWatched: film['user_details']['already_watched'],
-        watchingDate: film['user_details']['watching_date'],
-      }
-    };
+  updateFilmOnClient(updateType, updatedFilm) {
+    const index = this.#films.findIndex((film) => film.id === updatedFilm.id);
 
-    delete adaptedFilm['filmInfo']['age_rating'];
-    delete adaptedFilm['filmInfo']['alternative_title'];
-    delete adaptedFilm['filmInfo']['total_rating'];
-    delete adaptedFilm['userDetails']['already_watched'];
-    delete adaptedFilm['userDetails']['watching_date'];
+    if (index === -1) {
+      throw new Error('Can\'t update unexisting film');
+    }
 
-    return adaptedFilm;
+    this.#films = [
+      ...this.#films.slice(0, index),
+      updatedFilm,
+      ...this.#films.slice(index + 1),
+    ];
+
+    this._notify(updateType, updatedFilm);
   }
 
 }
