+++ src/service/adapter.js
@@ -1,0 +1,62 @@
+const adaptToServer = (film) => {
+  const adaptedFilm = {
+    'id': film.id,
+    'comments': film.comments,
+    'film_info': {
+      ...film.filmInfo,
+      'age_rating': film.filmInfo.ageRating,
+      'alternative_title': film.filmInfo.alternativeTitle,
+      'total_rating': film.filmInfo.ageRating,
+      'release': {
+        'date': film.filmInfo.release.date,
+        'release_country': film.filmInfo.release.releaseCountry
+      }
+    },
+    'user_details': {
+      ...film.userDetails,
+      'already_watched': film.userDetails.alreadyWatched,
+      'watching_date': film.userDetails.watchingDate,
+    }
+  };
+
+  delete adaptedFilm.film_info.ageRating;
+  delete adaptedFilm.film_info.alternativeTitle;
+  delete adaptedFilm.film_info.totalRating;
+  delete adaptedFilm.user_details.alreadyWatched;
+  delete adaptedFilm.user_details.watchingDate;
+
+  return adaptedFilm;
+};
+
+const adaptToClient = (film) => {
+  const adaptedFilm = {
+    id: film['id'],
+    comments: film['comments'],
+    filmInfo: {
+      ...film['film_info'],
+      ageRating: film['film_info']['age_rating'],
+      alternativeTitle: film['film_info']['alternative_title'],
+      totalRating: film['film_info']['total_rating'],
+      release: {
+        date: film['film_info']['release']['date'],
+        releaseCountry: film['film_info']['release']['release_country']
+      }
+    },
+    userDetails: {
+      ...film['user_details'],
+      alreadyWatched: film['user_details']['already_watched'],
+      watchingDate: film['user_details']['watching_date'],
+    }
+  };
+
+  delete adaptedFilm['filmInfo']['age_rating'];
+  delete adaptedFilm['filmInfo']['alternative_title'];
+  delete adaptedFilm['filmInfo']['total_rating'];
+  delete adaptedFilm['userDetails']['already_watched'];
+  delete adaptedFilm['userDetails']['watching_date'];
+
+  return adaptedFilm;
+};
+
+export { adaptToServer, adaptToClient };
+
