+++ src/service/api/films-api-service.js
@@ -1,0 +1,28 @@
+import ApiService from '../../framework/api-service.js';
+import { adaptToServer } from '../adapter.js';
+
+const Method = {
+  GET: 'GET',
+  PUT: 'PUT',
+};
+
+export default class FilmsApiService extends ApiService {
+  get films() {
+    return this._load({url: 'movies'})
+      .then(ApiService.parseResponse);
+  }
+
+  async updateFilm(film) {
+    const response = await this._load({
+      url: `movies/${film.id}`,
+      method: Method.PUT,
+      body: JSON.stringify(adaptToServer(film)),
+      headers: new Headers({'Content-Type': 'application/json'}),
+    });
+
+    const parsedResponse = await ApiService.parseResponse(response);
+
+    return parsedResponse;
+  }
+
+}
